put the sslbypass in a new class and parses drugs based on the groups above them
